% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmake.R
\name{as_data_file}
\alias{as_data_file}
\title{Return the data file name corresponding to the given indicator name}
\usage{
as_data_file(indicator)
}
\arguments{
\item{ind_file}{the indicator name (with path as needed) whose corresponding
data file name should be returned}
}
\description{
If `ind_file` does not have the indicator extension, an error will be
generated.
}
