% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmake.R
\name{as_ind_file}
\alias{as_ind_file}
\title{Returns the indicator name corresponding to the given data file name}
\usage{
as_ind_file(data_file, ind_ext = getOption("scipiper.ind_ext"))
}
\arguments{
\item{data_file}{the data file name (with path as needed) whose corresponding
indicator name should be returned}

\item{ind_ext}{the indicator file extension to apply}
}
\description{
If \code{data_file} already has the indicator extension, an error will be
generated.
}
\examples{
as_ind_file('mydata.rds') # 'mydata.rds.ind'
as_ind_file('mydata.rds', ind_ext='st') # 'mydata.rds.st'
as_ind_file('mydata.rds.ind') # Error: "data_file is an indicator file already"
}
