% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmake.R
\name{as_indicator}
\alias{as_indicator}
\title{Returns the indicator name corresponding to the given data file name}
\usage{
as_indicator(data_file)
}
\arguments{
\item{data_file}{the data file name (with path as needed) whose corresponding
indicator name should be returned}
}
\description{
If \code{data_file} already has the indicator extension, an error will be
generated.
}
