% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{check_local_source}
\alias{check_local_source}
\title{decide whether local_source is an indicator or data file and find the data file
if it is an indicator}
\usage{
check_local_source(local_source, ind_ext)
}
\description{
decide whether local_source is an indicator or data file and find the data file
if it is an indicator
}
\keyword{internal}
