% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmake.R
\name{combine_to_ind}
\alias{combine_to_ind}
\title{Create an indicator file that contains hashes of other files}
\usage{
combine_to_ind(ind_file, ...)
}
\arguments{
\item{ind_file}{file name of the indicator file to write}

\item{...}{files to combine into a single indicator file}
}
\description{
Create an indicator file that contains hashes of other files
}
\details{
light wrapper on \code{sc_indicate}
}
