% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmake.R
\name{combine_to_tibble}
\alias{combine_to_tibble}
\title{Create an object that contains hashes of other files}
\usage{
combine_to_tibble(...)
}
\arguments{
\item{...}{files to combine into a single indicator file}
}
\description{
Create an object that contains hashes of other files
}
\details{
light wrapper on \code{sc_indicate}
}
\examples{
tfiles <- tempfile(pattern=as.character(1:3))
file.create(tfiles)
combine_to_tibble(tfiles[1])
combine_to_tibble(tfiles[1:2])
combine_to_tibble(tfiles[1:3])
}
