% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_task_makefile.R
\name{create_task_table}
\alias{create_task_table}
\title{Convert a task_plan into a status table}
\usage{
create_task_table(task_plan, table_file)
}
\arguments{
\item{task_plan}{a task plan as produced by `create_task_plan()`}

\item{table_file}{a file name to write a tab-separated table to (as .tsv), or
NULL to return as a data.frame}
}
\description{
Create a data.frame or .tsv file representing a task_plan and our status
within that plan
}
