% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_task_makefile.R
\name{evaluate_step_element}
\alias{evaluate_step_element}
\title{Evaluate a step element into simple strings}
\usage{
evaluate_step_element(task_step, element, task, step)
}
\arguments{
\item{task_step}{a task_step as created by `create_task_step()`}

\item{element}{character naming a task_step element ('target_name',
'depends', or 'command') to evaluate}

\item{task}{a task item: a list with fields for 'task_name' and possibly
other info}

\item{step}{a step item: a list with fields for 'step_name' and possibly
other info}
}
\description{
Evaluate a step element (target, depends, command) as a character
string/vector or as a function of the task and step elements (depending on
the class of the element)
}
\keyword{internal}
