% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remake.R
\name{export_remake_status}
\alias{export_remake_status}
\title{Export status files from .remake folder (binary form) to status folder
(versionable text)}
\usage{
export_remake_status()
}
\description{
Export status files from .remake folder (binary form) to status folder
(versionable text)
}
