% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmake.R
\name{fetch}
\alias{fetch}
\title{Fetch objects from the scipiper data store without rebuilding them}
\usage{
fetch(
  target_name,
  remake_file = getOption("scipiper.remake_file"),
  require_current = FALSE,
  verbose = TRUE,
  allow_missing_packages = FALSE
)
}
\description{
Identical to \code{remake::fetch()} except that \code{remake_file} is the second argument
}
