% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{find_local_file}
\alias{find_local_file}
\title{decide whether local_source is an indicator or data file and find the data file
if it is an indicator}
\usage{
find_local_file(local_source, ind_ext)
}
\description{
decide whether local_source is an indicator or data file and find the data file
if it is an indicator
}
\keyword{internal}
