% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_drive.R
\name{gd_config}
\alias{gd_config}
\title{Create or overwrite a Google Drive configuration file}
\usage{
gd_config(folder, config_file = options("scipiper.gd_config_file")[[1]])
}
\arguments{
\item{folder}{character name of the Google Drive folder where all files for
this project are to be stored}

\item{config_file}{character name of the yml file where this configuration
information should be written}
}
\description{
Create or overwrite a Google Drive configuration file
}
