% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_drive.R
\name{gd_confirm_posted}
\alias{gd_confirm_posted}
\title{Check whether a file is on Google Drive, and if so, write an indicator file}
\usage{
gd_confirm_posted(ind_file,
  config_file = options("scipiper.gd_config_file")[[1]])
}
\arguments{
\item{ind_file}{character name of the indicator file to write locally once
the file has been uploaded; will exactly correspond to the data_file on GD}

\item{config_file}{character name of the YAML file containing
project-specific configuration information}
}
\description{
Check whether a file is on Google Drive, and if so, write an indicator file
}
