% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_drive.R
\name{gd_make_indicator}
\alias{gd_make_indicator}
\title{Write an Google Drive indicator file}
\usage{
gd_make_indicator(ind_file, md5_checksum)
}
\arguments{
\item{ind_file}{character name of the indicator file to write locally once
the file has been uploaded}

\item{config_file}{character name of the YAML file containing project-specific
configuration information}
}
\description{
Write an indicator file using a standard format for gd files
}
\keyword{internal}
