% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_drive.R
\name{gd_read_time}
\alias{gd_read_time}
\title{Reads a datetime as returned from AWS Google Drive}
\usage{
gd_read_time(datetime)
}
\arguments{
\item{datetime}{character datetime as from an aws.gd call}
}
\value{
POSIXct datetime
}
\description{
Reads in an AWS Google Drive standard datetime into POSIXct
}
\keyword{internal}
