% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_info.R
\name{get_dependency_status}
\alias{get_dependency_status}
\title{Retrieve status of dependencies from the remake store}
\usage{
get_dependency_status(
  target_name,
  remake_object,
  as_of = c("last_build", "now"),
  format = c("raw", "wide", "long")
)
}
\arguments{
\item{target_name}{length-1 name of a target}

\item{remake_object}{as produced by remake:::remake}

\item{as_of}{what status should be reflected - status as of the last build
(from the remake database) or status if we considered it built right now?}

\item{format}{return in remake's list format ('raw'), a one-row tibble with
nested tibbles for depends and code ('wide'), or a multi-row tibble with
columns for name, dependency type, and dependency hash ('long')?}
}
\description{
Returns placeholders if the object doesn't yet exist. hash=NA when the target
hasn't yet been built, or '??' when the target is fake, or a hexcode string
otherwise.
}
