% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remake.R
\name{get_mangled_key}
\alias{get_mangled_key}
\title{Convert keys into mangled keys as used in the .remake storr}
\usage{
get_mangled_key(key, dbstore = remake:::remake(load_sources = FALSE)$store$db)
}
\arguments{
\item{key}{character vector of key[s] to convert}

\item{dbstore}{a storr containing the remake build status}
}
\description{
Whereas file and object names might be invalid or confusing as key file
names, mangled keys are always good as file names - no punctuation or
misleading suffixes
}
\keyword{internal}
