% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_targets.R
\name{hash_dependencies}
\alias{hash_dependencies}
\title{Hash the depends objects of a target}
\usage{
hash_dependencies(target_name, remake_file)
}
\arguments{
\item{target_name}{the name of the single target whose dependencies should be
hashed}

\item{remake_file}{the name of the remake file where \code{target_name}'s recipe
is given}
}
\description{
Returns a named vector of hashes for all the dependencies of the target
}
