% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_targets.R
\name{hash_targets}
\alias{hash_targets}
\title{Create a vector of hashes corresponding to a vector of targets}
\usage{
hash_targets(target_names, target_is_file, file_extensions = NULL)
}
\arguments{
\item{target_names}{character vector of target names}

\item{target_is_file}{logical vector, same length as target_names, with TRUE
for targets that are files and FALSE for targets that are objects}

\item{file_extensions}{if target_is_file is not specified, file_extensions
can be used to supplement the default remake::file_extensions() to infer
which targets are files from the target_names. If file_extensions=NULL,
only the remake defaults will be used}
}
\description{
Create a vector of hashes corresponding to a vector of targets
}
