% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_drive.R
\name{indicate}
\alias{indicate}
\title{Write an indicator file}
\usage{
indicate(ind_file, md5_checksum)
}
\arguments{
\item{ind_file}{character name of the indicator file to write locally once
the file has been uploaded}

\item{md5_checksum}{character md5 hash}
}
\description{
Write an indicator file using a standard format
}
\keyword{internal}
