% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmake.R
\name{indicator_extension}
\alias{indicator_extension}
\title{Returns the indicator file extension for this project}
\usage{
indicator_extension(remake_file = options("scipiper.remake_file")[[1]])
}
\arguments{
\item{remake_file}{name of the remake YAML file}
}
\description{
The default extension is 'ind', but you can override this by defining a
remake recipe for \code{indicator_extension} in your remake_file, where the result
of that recipe is a length-1 character extension (with no leading period)
}
