% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmake.R
\name{is_indicator}
\alias{is_indicator}
\title{Determine whether target_names are status indicator files}
\usage{
is_indicator(target_names)
}
\arguments{
\item{target_names}{character vector of remake target names}

\item{remake_file}{filename of the remake YAML file}
}
\description{
Status indicator files are those files (or maybe someday objects?) included
in the remake yml whose final extension is the accepted indicator extension
('ind' by default, but see \code{indicator_extension}). If the target does not
have the indicator extension, FALSE is returned; no warnings or errors are
given if the target is not in the remake yml.
}
\details{
By default, the only accepted indicator extension is 'ind'. If you want other
extensions to be used, add a object target to your remake.yml that contains a
character vector of the accepted extensions. See below for an example.
}
\examples{
\dontrun{
# example remake.yml target to define extensions
targets:
  indicator_extensions:
    command: c(I("ind"))
}
}
