% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmake.R
\name{is_status_indicator}
\alias{is_status_indicator}
\title{Determine whether target_names are status indicator files}
\usage{
is_status_indicator(target_names, remake_file = "remake.yml")
}
\description{
Status indicator files are those files (not objects) included in the remake
yml whose final extension is one of the accepted indicator file extensions
('st' by default, but see Details). If any of these criteria is not met,
FALSE is returned; no warnings or errors are given if the target is not in
the remake yml.
}
\details{
By default, the only accepted indicator file extension is 'st'. If you want
other extensions to be used, add a object target to your remake.yml that
contains a character vector of the accepted extensions. See below for a
yaml-based example.
}
\examples{
\dontrun{
# example remake.yml target to define extensions
targets:
  indicator_file_extensions:
    packages: yaml
    command: yaml.load(I("[st,s3,yeti]"))
}
}
