% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_info.R
\name{list_all_targets}
\alias{list_all_targets}
\title{Get a list of all targets in a remake file}
\usage{
list_all_targets(
  remake_file = getOption("scipiper.remake_file"),
  recursive = FALSE
)
}
\arguments{
\item{remake_file}{filename of the remake YAML file from which targets should
be collected}

\item{recursive}{logical. if TRUE, result will include all targets from any
YAMLs listed in the include: section of the given remake_file, or any YAMLs
listed in the include: sections of those included YAMLs, etc.}
}
\value{
vector of all target names explicitly declared in this remake_file
(and if recursive=TRUE, also the names of targets declared in remake files
included by this remake_file)
}
\description{
Get a list of all targets in a remake file
}
\examples{
\dontrun{
# assuming you have a file named remake.yml:
list_all_targets() # get status for all explicitly named targets in remake.yml

# status for all targets in a different remake YAML:
list_all_targets('other_remake.yml')

# status for all targets in remake.yml and any remake YAMLs included by remake.yml
list_all_targets(recursive=TRUE)
}
}
