% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmake.R
\name{list_group_targets}
\alias{list_group_targets}
\title{list targets in remake file that are fake/dummy/group targets}
\usage{
list_group_targets(remake_file = getOption("scipiper.remake_file"))
}
\arguments{
\item{remake_file}{filename of the remake YAML file from which targets should
be collected}
}
\description{
list targets in remake file that are fake/dummy/group targets
}
\details{
this is an internal file. We may want to make it recursive, but I
don't see a use-case right now for that.
}
\keyword{internal}
