% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-scipiper.R
\docType{package}
\name{options}
\alias{options}
\title{Scipiper session options}
\description{
Several session-level options are stored in R's \code{.Options} object.
}
\details{
\describe{

\item{\code{scipiper.remake_file}}{File path of the main remake YAML file for a
project. Used as the default for the \code{remake_file} argument in many
functions.}

\item{\code{scipiper.dry_put}}{Logical, default FALSE. If TRUE, calls to \code{s3_put} or \code{gd_put} won't actually push anything to S3 or Google Drive; they'll just pretend they've done it.}

\item{\code{scipiper.s3_config_file}}{File path to a configuration file for an
Amazon S3 bucket. Used as the default for the \code{s3_config_file} argument in
many functions (usually those prefixed with \code{s3_}).}

\item{\code{scipiper.gd_config_file}}{File path to a configuration file for a
Google Drive folder. Used as the default for the \code{gd_config_file} argument in
many functions (usually those prefixed with \code{gd_}).}

\item{\code{scipiper.ind_ext}}{The indicator file extension to recognize, i.e.,
the final file extension of files for which \code{is_ind_file()} should return
\code{TRUE} and which \code{as_ind_file} should append to the data file name. Used as
the default for the \code{ind_ext} argument in many functions.`}

}
}
\examples{
# see what's already there
getOption('scipiper.remake_file') # 'remake.yml'
getOption('scipiper.gd_config_file') # "lib/cfg/gd_config.yml"
getOption('scipiper.s3_config_file') # "lib/cfg/s3_config.yml"
getOption('scipiper.ind_ext') #'ind'
.Options[grep('scipiper', names(.Options), value=TRUE)] # list all scipiper options

# change the options - these commands can be placed in your project's .Rprofile
# file if you want them to be reapplied in every R session for the project
options('scipiper.ind_ext'='st', 'scipiper.remake_file'='remake.yaml')
}
