% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamps.R
\name{POSIX2char}
\alias{POSIX2char}
\title{Format a POSIXct timestamp with UTC into a character string}
\usage{
POSIX2char(ptime)
}
\description{
Format a POSIXct timestamp with UTC into a character string
}
