% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require_libs.R
\name{require_libs}
\alias{require_libs}
\title{Ensure the listed libraries are available}
\usage{
require_libs(...)
}
\arguments{
\item{...}{character-string names of packages to require. argument names are
ignored.}
}
\description{
Calls requireNamespace for the packages listed in ..., throwing errors if any
package is unavailable.
}
