% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retry_patiently.R
\name{retry_patiently}
\alias{retry_patiently}
\title{Evaluate an expression with retries, waiting between tries}
\usage{
retry_patiently(expr, max_wait_times = c(0, 3^c(0:4)),
  min_wait_frac = 0.5, verbose = FALSE)
}
\arguments{
\item{expr}{expression to evaluate}

\item{max_wait_times}{vector of maximum wait times, in seconds, before trying
the 1st, 2nd, etc. time. The length of this vector is the maximum number of
tries.}

\item{min_wait_frac}{numeric between 0 and 1. If < 1, the actual wait time
before try \code{i} will be a uniform random value selected from the range
\code{(min_wait_frac*max_wait_times[i], max_wait_times[i])}. Modeled loosely on
https://github.com/r-lib/httr/blob/master/R/retry.R.}

\item{verbose}{logical. Should informational messages be printed?}
}
\description{
Waits progressively longer before each next try (at least by default; wait
times can be customized).
}
