% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_confirm_posted}
\alias{s3_confirm_posted}
\title{Check whether a file is on S3, and if so, write an indicator file}
\usage{
s3_confirm_posted(ind_file,
  config_file = getOption("scipiper.s3_config_file"),
  ind_ext = getOption("scipiper.ind_ext"))
}
\arguments{
\item{ind_file}{character name of the indicator file to write locally once
the file has been uploaded}

\item{config_file}{character name of the yml file containing project-specific
configuration information}

\item{ind_ext}{the indicator file extension to expect at the end of ind_file}
}
\description{
Check whether a file is on S3, and if so, write an indicator file
}
