% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_list}
\alias{s3_list}
\title{List the S3 objects for this project}
\usage{
s3_list(..., config_file = getOption("scipiper.s3_config_file"))
}
\arguments{
\item{...}{arguments passed to aws.s3::get_bucket_df}

\item{config_file}{character name of the yml file containing project-specific
configuration information}
}
\description{
List the S3 objects in the project bucket/path as given in config_file
}
