% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_make_indicator}
\alias{s3_make_indicator}
\title{Write an S3 indicator file}
\usage{
s3_make_indicator(ind_file, remote_time)
}
\arguments{
\item{ind_file}{character name of the indicator file to write locally once
the file has been uploaded}

\item{config_file}{character name of the yml file containing project-specific
configuration information}
}
\description{
Write an indicator file using a standard format for s3 files
}
\keyword{internal}
