% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_read_time}
\alias{s3_read_time}
\title{Reads a datetime as returned from AWS S3}
\usage{
s3_read_time(datetime)
}
\arguments{
\item{datetime}{character datetime as from an aws.s3 call}
}
\value{
POSIXct datetime
}
\description{
Reads in an AWS S3 standard datetime into POSIXct
}
\keyword{internal}
