% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmake.R
\name{scbless}
\alias{scbless}
\alias{sc_declare_current}
\title{Declare to remake and scipiper that a target is current}
\usage{
scbless(target_names, remake_file = getOption("scipiper.remake_file"),
  ind_ext = getOption("scipiper.ind_ext"))
}
\arguments{
\item{target_names}{names of the targets to record as already current}

\item{remake_file}{the file path+name of the remake file to use in building
the targets. As in remake::make, except that for scmake this param comes
before ... and second in line, so it can be easily specified without naming
the argument}

\item{ind_ext}{the indicator file extension identifying those files for which
build/status information will be shared via git-committable files in the
build/status folder. You should git commit the resulting build/status
files.}
}
\description{
\code{scbless} (aka \code{sc_declare_current}) tells remake and scipiper to update
their databases to reflect that the current target is up to date with respect
to the current dependencies
}
