% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-scipiper.R
\docType{package}
\name{scipiper}
\alias{scipiper}
\alias{scipiper-package}
\title{Scipiper package: Streamlining scientific data analysis workflows}
\description{
Scipiper provides support for:
}
\details{
\itemize{

\item Storing some project files in a shared S3 bucket or Google Drive folder

\item Combining \code{remake} with a shared remote cache, e.g., shared files in an
S3 bucket or a Google Drive folder

\item Planning and executing long lists (batches) of similar tasks, e.g.,
downloading data for all 50 states or running many models at many sites.
Scipiper includes fault tolerance and the ability to pick up partway through
a long batch run

}

See ?scipiper::options for global (session-level) options you can set for use
with scipiper.
}
