% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_dirs}
\alias{setup_dirs}
\title{Create standard project directories}
\usage{
setup_dirs(..., default = "cos", .list = list(`1_data` = default, `9_report`
  = default, lib = ""))
}
\arguments{
\item{...}{named characters where name=directory_name and
value=character_codes_for_dirs, or unnamed characters where
value=directory_name (and \code{default} will be used for
character_codes_for_dirs). Permitted letters for character_codes_for_dirs
are c=cfg, d=doc, j=job, o=out, and s=src}

\item{default}{string identifying the default set of
character_codes_for_dirs. one letter per subdirectory. Permitted letters
are c=cfg, d=doc, j=job, o=out, and s=src}

\item{.list}{as an alternative to ..., use .list to specify directories in a
single list argument}
}
\description{
Create standard project directories
}
\examples{
# prep
dir.create('example_project')
setwd('example_project')

setup_dirs()
setup_dirs('2_data','3_analyze','4_report','explore'='',default='cdos')
setup_dirs('1a_siteinfo','1b_discharge','2_clean'='dos','3_model'='cjs','explore'='')

# clean up
setwd('..')
unlink('example_project', recursive=TRUE)
}
