% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_files}
\alias{setup_files}
\title{Create standard project directories}
\usage{
setup_files(
  ...,
  .list = list(file.path(grep("/out$", list.dirs("."), value = TRUE), "README.md"), . =
    c("README.md", "Makefile"))
)
}
\arguments{
\item{...}{named characters where name=directory_name and value=filename\link{s},
or unnamed characters where value=filepath\link{s}. Paths can be relative.}

\item{.list}{as an alternative to ..., use .list to specify files in a single
list argument}
}
\description{
Create standard project directories
}
\examples{
# prep
dir.create('example_project')
setwd('example_project')

dirs <- setup_dirs('2_data','3_analyze','4_report','explore',default='cdos')
setup_files() # use the default as given for .list
setup_files('2_data/src'=c('data1.R','data2.R','helper-data.R'))

# clean up
setwd('..')
unlink('example_project', recursive=TRUE)
}
