% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamps.R
\name{time_conversions}
\alias{time_conversions}
\alias{POSIX2char}
\alias{char2POSIX}
\title{Format a POSIXct timestamp with UTC into a character string}
\usage{
POSIX2char(ptime)

char2POSIX(stime)
}
\arguments{
\item{ptime}{time in POSIXct format}

\item{stime}{time in character string format}
}
\value{
time in character string format

time in POSIXct format
}
\description{
Format a POSIXct timestamp with UTC into a character string

Format a character string into POSIXct timestamp with UTC
}
