% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_info.R
\name{which_dirty}
\alias{which_dirty}
\title{List the targets for which is_current is FALSE (as a character vector)}
\usage{
which_dirty(target_names = NULL,
  remake_file = getOption("scipiper.remake_file"), RDSify_first = TRUE)
}
\arguments{
\item{target_names}{character vector of targets for which to determine build
status, including status for dependencies of the named targets. If NULL
will return status for the default target and its dependencies.}

\item{remake_file}{filename of the remake YAML file from which status should
be determined}

\item{RDSify_first}{logical. Should the info in build/status/*.yml files be
copied over to the remake RDS-based status database before querying for
target cleanliness?}
}
\description{
Runs faster when RDSify_first = FALSE.
}
\examples{
\dontrun{
# assuming you have a file named remake.yml:
get_remake_status() # get status for the default target and its dependencies
get_remake_status(list_all_targets()) # get status for all explicitly named targets in remake.yml

# or to get status for all targets in a different remake YAML:
get_remake_status(list_all_targets('other_remake.yml'), 'other_remake.yml')
}
}
