% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamps.R
\name{write_indicator}
\alias{write_indicator}
\title{Write a timestamp to an indicator file}
\usage{
write_indicator(indicator_file, time = Sys.time())
}
\arguments{
\item{indicator_file}{character name/path of indicator file where the
timestamp should be written (overwrites existing)}

\item{time}{POSIXct time to write}
}
\description{
Write a timestamp to an indicator file
}
