% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remake.R
\name{YAMLify_build_status}
\alias{YAMLify_build_status}
\title{Copy info from .remake/objects to build/status}
\usage{
YAMLify_build_status(target_names)
}
\description{
Copy status files from .remake folder (binary form) to build/status folder
(versionable text)
}
\keyword{internal}
