% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_drive.R
\name{gd_confirm_posted}
\alias{gd_confirm_posted}
\title{Check whether a file is on Google Drive, and if so, write an indicator file}
\usage{
gd_confirm_posted(ind_file,
  config_file = getOption("scipiper.gd_config_file"),
  ind_ext = getOption("scipiper.ind_ext"))
}
\arguments{
\item{ind_file}{character name of the indicator file to write locally once
the file has been uploaded; will exactly correspond to the data_file on GD}

\item{config_file}{character name of the YAML file containing
project-specific configuration information}

\item{ind_ext}{the indicator file extension to expect at the end of ind_file}
}
\description{
Check whether a file is on Google Drive, and if so, write an indicator file
}
