% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_drive.R
\name{gd_confirm_posted}
\alias{gd_confirm_posted}
\title{Check whether a file is on Google Drive, and if so, write an indicator file}
\usage{
gd_confirm_posted(data_file, ind_file,
  config_file = options("scipiper.gd_config_file")[[1]])
}
\arguments{
\item{data_file}{character name of the data file to download (downloads the
Google Drive object whose key equals the data_file basename)}

\item{ind_file}{character name of the indicator file to write locally once
the file has been uploaded}

\item{config_file}{character name of the yml file containing project-specific
configuration information}
}
\description{
Check whether a file is on Google Drive, and if so, write an indicator file
}
