% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_drive.R
\name{gd_get}
\alias{gd_get}
\title{Download a file from Google Drive if needed}
\usage{
gd_get(ind_file, type = NULL, overwrite = TRUE, verbose = FALSE,
  config_file = getOption("scipiper.gd_config_file"),
  ind_ext = getOption("scipiper.ind_ext"))
}
\arguments{
\item{ind_file}{character name of the indicator file for which data should be
downloaded. downloads the Google Drive object whose key equals the
data_file basename}

\item{type}{see \code{type} argument to \code{googledrive::drive_download()}}

\item{overwrite}{see \code{overwrite} argument to \code{googledrive::drive_download()}}

\item{verbose}{see \code{verbose} argument to \code{googledrive::drive_download()};
also used to determine whether to include messages specific to \code{gd_get()}}

\item{config_file}{character name of the YAML file containing
project-specific configuration information for Google Drive}

\item{ind_ext}{the indicator file extension to expect at the end of ind_file}
}
\description{
Download a file from Google Drive to the local project based on the
information implied by the indicator file (including the location on google
drive and the local destination location). Skips the download if the local
file already exists and the remote and local hashes are identical.
}
\examples{
\dontrun{
gd_get('0_test/test_sheet.ind', type='xlsx', overwrite=TRUE)
}
}
