% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_drive.R
\name{gd_list}
\alias{gd_list}
\title{List the Google Drive objects for this project}
\usage{
gd_list(..., config_file = options("scipiper.gd_config_file")[[1]])
}
\arguments{
\item{...}{arguments passed to googledrive::drive_ls}

\item{config_file}{character name of the yml file containing project-specific
configuration information}
}
\description{
List the Google Drive objects in the project bucket/path as given in config_file
}
