% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_drive.R
\name{gd_put}
\alias{gd_put}
\title{Upload a file to Google Drive}
\usage{
gd_put(data_file, ind_file = as_indicator(data_file),
  on_exists = c("update", "replace", "stop"), type = NULL,
  verbose = FALSE, config_file = options("scipiper.gd_config_file")[[1]])
}
\arguments{
\item{data_file}{character name of the data file to upload. The basename will
be used as the Google Drive key and must be unique within the project
(bucket & path)}

\item{ind_file}{character name of the indicator file to write locally once
the file has been uploaded}

\item{on_exists}{what to do if the file already exists - update, replace, or
throw an error? the default is to update (using google drive's versioning
functionality)}

\item{type}{media type as passed to drive_upload or drive_update}

\item{verbose}{logical, used in gd_put and passed onto
googledrive::drive_update, drive_upload, and/or drive_rm}

\item{config_file}{character name of the yml file containing project-specific
configuration information}
}
\description{
Upload (create or overwrite) a file to the project bucket and path.
}
