% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_drive.R
\name{gd_put}
\alias{gd_put}
\title{Upload a file to Google Drive}
\usage{
gd_put(data_file, ind_file, config_file = "lib/cfg/gd_config.yml")
}
\arguments{
\item{data_file}{character name of the data file to upload. The basename will
be used as the Google Drive key and must be unique within the project (bucket & path)}

\item{ind_file}{character name of the indicator file to write locally once
the file has been uploaded}

\item{config_file}{character name of the yml file containing project-specific
configuration information}
}
\description{
Upload (create or overwrite) a file to the project bucket and path.
}
