% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmake.R
\name{get_remake_status}
\alias{get_remake_status}
\title{Produce a table describing the remake build status relative to 1+ targets}
\usage{
get_remake_status(target_names, remake_file = "remake.yml")
}
\arguments{
\item{target_names}{character vector of targets for which to determine build
status (complete status will include dependencies of these targets)}

\item{remake_file}{filename of the remake YAML file}
}
\description{
Produce a table describing the remake build status relative to 1+ targets
}
