% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_info.R
\name{get_remake_status}
\alias{get_remake_status}
\title{Produce a table describing the remake build status relative to 1+ targets}
\usage{
get_remake_status(
  target_names = NULL,
  remake_file = getOption("scipiper.remake_file"),
  format = c("nested", "flat"),
  RDSify_first = TRUE
)
}
\arguments{
\item{target_names}{character vector of targets for which to determine build
status, including status for dependencies of the named targets. If NULL
will return status for the default target and its dependencies.}

\item{remake_file}{filename of the remake YAML file from which status should
be determined}

\item{format}{if "nested", returnes a nested tibble with tibble list columns
for 'depends' and 'functions'. if "flat", returns character columns for
'depends' and 'functions' that are ugly, flattened versions of the tibbles.
"flat" can be useful for comparing status before and after an operation
using dplyr joins (e.g., within \code{scmake()}).}

\item{RDSify_first}{logical. Should the info in build/status/*.yml files be
copied over to the remake RDS-based status database before querying for
target cleanliness?}
}
\description{
Describes the hash, build time, and dependencies as of the most recent build
of each target. If the dependencies have changed since that last build, those
new dependencies are not described here. See why_dirty() to compare the old
and new dependencies.
}
\details{
Runs faster when RDSify_first = FALSE, but the default is TRUE so that you
can usually pretend that there's no extra step between a git pull and
updating the remake database.
}
\examples{
\dontrun{
# assuming you have a file named remake.yml:
get_remake_status() # get status for the default target and its dependencies
get_remake_status(list_all_targets(), RDSify_first=FALSE) # status for all targets in remake.yml

# or to get status for all targets in a different remake YAML:
get_remake_status(list_all_targets('other_remake.yml'), 'other_remake.yml')
}
}
