% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psprintf.R
\name{psprintf}
\alias{psprintf}
\title{Paste together sprintf-style text and variable values}
\usage{
psprintf(..., sep = "\\n      ")
}
\arguments{
\item{...}{list of sprintf arguments seperated by commas}

\item{sep}{separator for the list of arguments, defaults to new line}
}
\description{
Helper function to sprintf a bunch of key-value (string-variableVector) pairs,
then paste them together with a good separator for contructing remake recipes.
}
\examples{
\dontrun{
psprintf(
  "plot_frame(",
  "png_file = target_name,",
  "date = date_\%s," = "2018-01-23",
  "state = \%s)" = "Wisconsin")
}
}
