% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_config}
\alias{s3_config}
\title{Create or overwrite an S3 configuration file}
\usage{
s3_config(
  bucket,
  profile = "default",
  config_file = getOption("scipiper.s3_config_file")
)
}
\arguments{
\item{bucket}{character name of an S3 bucket}

\item{profile}{character name of an S3 profile to use}

\item{config_file}{character name of the yml file where this configuration
information should be written}
}
\description{
It is assumed that the bucket corresponds 1:1 with the local project directory
structure, so all object keys will be the same as the local file path from the
root of the project directory.
}
