% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_confirm_posted}
\alias{s3_confirm_posted}
\title{Check whether a file is on S3, and if so, write an indicator file}
\usage{
s3_confirm_posted(data_file, ind_file, config_file = "lib/cfg/s3_config.yml")
}
\arguments{
\item{data_file}{character name of the data file to download (downloads the
S3 object whose key equals the data_file basename)}

\item{ind_file}{character name of the indicator file to write locally once
the file has been uploaded}

\item{config_file}{character name of the yml file containing project-specific
configuration information}
}
\description{
Check whether a file is on S3, and if so, write an indicator file
}
