% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_get}
\alias{s3_get}
\title{Download a file from S3}
\usage{
s3_get(data_file, config_file = "lib/cfg/s3_config.yml")
}
\arguments{
\item{data_file}{character name of the data file to download (downloads the
S3 object whose key equals the data_file basename)}

\item{config_file}{character name of the yml file containing project-specific
configuration information}
}
\description{
Download a file from S3 to the local project
}
