% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_put}
\alias{s3_put}
\title{Upload a file to S3}
\usage{
s3_put(data_file, config_file = "lib/cfg/s3_config.yml")
}
\arguments{
\item{data_file}{character name of the data file to upload. The basename will
be used as the S3 key and must be unique within the project (bucket & path)}

\item{config_file}{character name of the yml file containing project-specific
configuration information}
}
\description{
Upload (create or overwrite) a file to the project bucket and path.
}
