% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmake.R
\name{sc_indicate}
\alias{sc_indicate}
\title{Create an indicator file}
\usage{
sc_indicate(indicator, ..., data_file)
}
\arguments{
\item{indicator}{file name of the indicator file to write}

\item{...}{optional. named character strings/vectors to be written to the
indicator file. one good option is a pre-computed hash of the actual data
file (possibly retrieved as a hash from the remote cache). If you have the
data_file locally and don't yet have a hash, just specify the \code{data_file}
argument instead.}

\item{data_file}{optional. file name of the data file whose presence is being
indicated. if given, the hash of the data file will be included in the
indicator file as the \code{hash} element.}
}
\description{
If only the first argument (\code{indicator}) is given, the contents of the
indicator file change every time. To create an indicator file whose contents
are static, specify a fixed argument in \code{...}.
}
