% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmake.R
\name{sc_retrieve}
\alias{sc_retrieve}
\title{Retrieve the data file declared by an indicator}
\usage{
sc_retrieve(indicator)
}
\arguments{
\item{indicator}{the file path of the indicator}
}
\description{
Identifies the data file's name by removing the indicator extension, then
calls \code{scmake} to retrieve that file using a recipe given in the remake.yml
}
