% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmake.R
\name{scmake}
\alias{scmake}
\title{Wrapper for remake::make that permits cache sharing}
\usage{
scmake(target_names = NULL, ..., verbose = TRUE,
  allow_missing_packages = FALSE, remake_file = "remake.yml")
}
\arguments{
\item{target_names}{as in remake::make}

\item{...}{as in remake::make}

\item{verbose}{as in remake::make}

\item{allow_missing_packages}{as in remake::make}

\item{remake_file}{as in remake::make}
}
\description{
The name: sc=scipiper, sc=shared cache. This wrapper could eventually involve
other custom hooks for things to do before or after building a target
}
