% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_info.R
\name{why_dirty}
\alias{why_dirty}
\title{Explain why a target is dirty}
\usage{
why_dirty(target_name, remake_file = getOption("scipiper.remake_file"),
  RDSify_first = FALSE)
}
\arguments{
\item{remake_file}{filename of the remake YAML file from which status should
be determined}

\item{RDSify_first}{logical. Should the info in build/status/*.yml files be
copied over to the remake RDS-based status database before querying for
target cleanliness? Defaults to FALSE because if you're asking why, you've
probably already queried to determine that the target is dirty}

\item{target_names}{character vector of targets for which to determine build
status, including status for dependencies of the named targets. If NULL
will return status for the default target and its dependencies.}
}
\description{
Compares the dependency status of a target (1) as it currently appears in the
repo and (2) as documented in the remake database. Returns information about
the mismatches that are causing remake to consider this target dirty.
}
